chore: separate files in sprt/web
